Remove the flag ozone.ec.grpc.zerocopy.enabled, always have zero-copy enabled.
